requirements.txt
# Core dependencies
openai>=1.50.0
pyyaml>=6.0.1
rich>=13.7.0
python-dotenv>=1.0.0
requests>=2.31.0
